Preliminary refactor of onnx import module
